Add transition for lights (#21)
* Remove redundant project from csproj

* Add config of light transition

---------

+++ src/Extensions/LightExtensionMethods/LightExtensionMethods.cs
@@ -34,12 +34,12 @@ public static class LightExtensionMethods
         target.CallService("turn_on");
     }
     
-    public static void Toggle(this ILightEntityCore target, long? transition = null, long? brightnessPct = null)
+    public static void Toggle(this ILightEntityCore target, double? transition = null, long? brightnessPct = null)
     {
         target.CallService("toggle", new LightParameters { Transition = transition, BrightnessPct = brightnessPct});
     }
     
-    public static void TurnOn(this IEnumerable<ILightEntityCore> target, long? transition = null, long? brightnessPct = null)
+    public static void TurnOn(this IEnumerable<ILightEntityCore> target, double? transition = null, long? brightnessPct = null)
     {
         target.CallService("toggle", new LightParameters { Transition = transition, BrightnessPct = brightnessPct});
     }
